Move client libraries into separate module
