+++ Makefile
@@ -57,10 +57,10 @@ build: bin/proxy-agent bin/proxy-server bin/proxy-test-client
 bin/proxy-agent: bin cmd/agent/main.go proto/agent/agent.pb.go
 	GO111MODULE=on go build -o bin/proxy-agent cmd/agent/main.go
 
-bin/proxy-test-client: bin cmd/client/main.go proto/proxy.pb.go
+bin/proxy-test-client: bin cmd/client/main.go proto/agent/agent.pb.go
 	GO111MODULE=on go build -o bin/proxy-test-client cmd/client/main.go
 
-bin/proxy-server: bin cmd/proxy/main.go proto/agent/agent.pb.go proto/proxy.pb.go
+bin/proxy-server: bin cmd/proxy/main.go proto/agent/agent.pb.go proto/agent/agent.pb.go
 	GO111MODULE=on go build -o bin/proxy-server cmd/proxy/main.go
 
 ## --------------------------------------
@@ -73,17 +73,12 @@ bin/proxy-server: bin cmd/proxy/main.go proto/agent/agent.pb.go proto/proxy.pb.g
 ## --------------------------------------
 
 .PHONY: gen
-gen: proto/agent/agent.pb.go proto/proxy.pb.go mock_gen
+gen: proto/agent/agent.pb.go mock_gen
 
-proto/agent/agent.pb.go: proto/agent/agent.proto
-	protoc -I proto proto/agent/agent.proto --go_out=plugins=grpc:proto
-	cat hack/go-license-header.txt proto/agent/agent.pb.go > proto/agent/agent.licensed.go
-	mv proto/agent/agent.licensed.go proto/agent/agent.pb.go
-
-proto/proxy.pb.go: proto/proxy.proto
-	protoc -I proto proto/proxy.proto --go_out=plugins=grpc:proto
-	cat hack/go-license-header.txt proto/proxy.pb.go > proto/proxy.licensed.go
-	mv proto/proxy.licensed.go proto/proxy.pb.go
+proto/agent/agent.pb.go: konnectivity-client/proto/agent/agent.proto
+	protoc -I konnectivity-client/proto konnectivity-client/proto/agent/agent.proto --go_out=plugins=grpc:konnectivity-client/proto
+	cat hack/go-license-header.txt konnectivity-client/proto/agent/agent.pb.go > konnectivity-client/proto/agent/agent.licensed.go
+	mv konnectivity-client/proto/agent/agent.licensed.go konnectivity-client/proto/agent/agent.pb.go
 
 ## --------------------------------------
 ## Certs
@@ -155,7 +150,7 @@ docker-push/proxy-agent: docker-build/proxy-agent
 	${DOCKER_CMD} push ${AGENT_FULL_IMAGE}-$(ARCH):${TAG}
 
 .PHONY: docker-build/proxy-server
-docker-build/proxy-server: cmd/proxy/main.go proto/agent/agent.pb.go proto/proxy.pb.go
+docker-build/proxy-server: cmd/proxy/main.go proto/agent/agent.pb.go
 	@[ "${TAG}" ] || ( echo "TAG is not set"; exit 1 )
 	echo "Building proxy-server for ${ARCH}"
 	${DOCKER_CMD} build . --build-arg ARCH=$(ARCH) -f artifacts/images/server-build.Dockerfile -t ${SERVER_FULL_IMAGE}-$(ARCH):${TAG}
@@ -166,7 +161,7 @@ docker-push/proxy-server: docker-build/proxy-server
 	${DOCKER_CMD} push ${SERVER_FULL_IMAGE}-$(ARCH):${TAG}
 
 .PHONY: docker-build/proxy-test-client
-docker-build/proxy-test-client: cmd/client/main.go proto/agent/agent.pb.go proto/proxy.pb.go
+docker-build/proxy-test-client: cmd/client/main.go proto/agent/agent.pb.go
 	@[ "${TAG}" ] || ( echo "TAG is not set"; exit 1 )
 	echo "Building proxy-test-client for ${ARCH}"
 	${DOCKER_CMD} build . --build-arg ARCH=$(ARCH) -f artifacts/images/client-build.Dockerfile -t ${TEST_CLIENT_FULL_IMAGE}-$(ARCH):${TAG}
@@ -249,4 +244,4 @@ release-alias-tag: # Adds the tag to the last build tag. BASE_REF comes from the
 
 .PHONY: clean
 clean:
-	rm -rf proto/agent/agent.pb.go proto/proxy.pb.go easy-rsa.tar.gz easy-rsa-master cfssl cfssljson certs bin proto/agent/mocks
+	rm -rf proto/agent/agent.pb.go easy-rsa.tar.gz easy-rsa-master cfssl cfssljson certs bin proto/agent/mocks
