+++ konnectivity-client/go.mod
@@ -1,0 +1,9 @@
+module sigs.k8s.io/apiserver-network-proxy/konnectivity-client
+
+go 1.13
+
+require (
+	github.com/golang/protobuf v1.3.3
+	google.golang.org/grpc v1.27.0
+	k8s.io/klog v1.0.0
+)
