+++ konnectivity-client/pkg/client/client.go
@@ -27,7 +27,7 @@ import (
 
 	"google.golang.org/grpc"
 	"k8s.io/klog"
-	"sigs.k8s.io/apiserver-network-proxy/proto/agent"
+	"sigs.k8s.io/apiserver-network-proxy/konnectivity-client/proto/agent"
 )
 
 // Tunnel provides ability to dial a connection through a tunnel.
