+++ pkg/agent/agentclient/client_test.go
@@ -11,7 +11,7 @@ import (
 
 	"google.golang.org/grpc"
 	"k8s.io/klog"
-	"sigs.k8s.io/apiserver-network-proxy/proto/agent"
+	"sigs.k8s.io/apiserver-network-proxy/konnectivity-client/proto/agent"
 )
 
 func TestServeData_HTTP(t *testing.T) {
