+++ tests/concurrent_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"google.golang.org/grpc"
-	"sigs.k8s.io/apiserver-network-proxy/pkg/agent/client"
+	"sigs.k8s.io/apiserver-network-proxy/konnectivity-client/pkg/client"
 )
 
 func TestProxy_Concurrency(t *testing.T) {
