+++ tests/ha_proxy_server_test.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	"google.golang.org/grpc"
-	"sigs.k8s.io/apiserver-network-proxy/pkg/agent/client"
+	"sigs.k8s.io/apiserver-network-proxy/konnectivity-client/pkg/client"
 )
 
 type tcpLB struct {
